Add User Settings (#4)
* add user settings

* fix codestyle

* use getKey instead of id in user settings

* remove duplication and add abstract method

* remove empty line

* cache on set instead of on get

* cache user settings

* make commands code cleaner

* use lf instead of crlf

* use printer package

* clean repeated code

* update readme

* update readme

+++ src/SettingsManger.php
@@ -7,36 +7,42 @@ use Merodiro\Settings\Models\Setting;
 
 class SettingsManger
 {
+    public function cacheKey($key)
+    {
+        return config('settings.cache_prefix') . $key . '_global';
+    }
+
     public function all()
     {
-        return Setting::pluck('value', 'key');
+        return Setting::whereNull('owner_id')->pluck('value', 'key');
     }
 
     public function set($key, $value)
     {
-        Setting::updateOrCreate(['key' => $key], ['value' => $value]);
+        Setting::updateOrCreate(['key' => $key, 'owner_id' => null], ['value' => $value]);
     }
 
     public function get($key, $default = null)
     {
-        $cache_key = config('settings.cache_prefix') . $key;
-        $duration = config('settings.cache_duration');
+        $cache_key = static::cacheKey($key);
 
-        $value = Cache::remember($cache_key, $duration, function () use ($key) {
-            return Setting::where('key', $key)->pluck('value')->first();
-        });
+        if (Cache::has($cache_key)) {
+            return Cache::get($cache_key);
+        }
+
+        $value = Setting::where('key', $key)->whereNull('owner_id')->pluck('value')->first();
 
         return $value ? $value : $default;
     }
 
     public function forget($key)
     {
-        Setting::where('key', $key)->first()->delete();
+        Setting::where('key', $key)->whereNull('owner_id')->first()->delete();
     }
 
     public function flush()
     {
-        Setting::get()->each(function ($item) {
+        Setting::whereNull('owner_id')->each(function ($item) {
             $item->delete();
         });
     }
